requirements.txt
seaborn==0.8.1
pandas==0.23.1
jsonpickle==0.9.6
numpy==1.14.5
matplotlib==2.2.2
GitPython==2.1.11
typing==3.6.4
